+++ x/protocolpool/keeper/genesis.go
@@ -2,8 +2,11 @@ package keeper
 
 import (
 	"context"
+	"errors"
 	"fmt"
+	"time"
 
+	"cosmossdk.io/math"
 	"cosmossdk.io/x/protocolpool/types"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -49,8 +52,21 @@ func (k Keeper) InitGenesis(ctx context.Context, data *types.GenesisState) error
 		}
 	}
 
-	if err := k.ToDistribute.Set(ctx, data.ToDistribute); err != nil {
-		return fmt.Errorf("failed to set to distribute: %w", err)
+	if err := k.LastBalance.Set(ctx, data.LastBalance); err != nil {
+		return fmt.Errorf("failed to set last balance: %w", err)
+	}
+
+	totalToBeDistributed := math.ZeroInt()
+	for _, distribution := range data.Distributions {
+		totalToBeDistributed = totalToBeDistributed.Add(distribution.Amount)
+		if err := k.Distributions.Set(ctx, *distribution.Time, distribution.Amount); err != nil {
+			return fmt.Errorf("failed to set distribution: %w", err)
+		}
+	}
+
+	// sanity check to avoid trying to distribute more than what is available
+	if data.LastBalance.LT(totalToBeDistributed) {
+		return errors.New("total to be distributed is greater than the last balance")
 	}
 
 	return nil
@@ -96,7 +112,19 @@ func (k Keeper) ExportGenesis(ctx context.Context) (*types.GenesisState, error) 
 
 	genState := types.NewGenesisState(cf, budget)
 
-	genState.ToDistribute, err = k.ToDistribute.Get(ctx)
+	genState.LastBalance, err = k.LastBalance.Get(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	err = k.Distributions.Walk(ctx, nil, func(key time.Time, value math.Int) (stop bool, err error) {
+		genState.Distributions = append(genState.Distributions, &types.Distribution{
+			Time:   &key,
+			Amount: value,
+		})
+
+		return false, nil
+	})
 	if err != nil {
 		return nil, err
 	}